DDCNLS-6929-a refactored out Tax Summary Liability as per code review
