+++ app/models/PensionTaxRate.scala
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2019 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package models
+
+import play.api.libs.json._
+
+final case class PensionTaxRate(value: Double) {
+  val percentage: Int = (value * 100).toInt
+}
+
+object PensionTaxRate {
+
+  implicit val reads: Reads[PensionTaxRate] = new Reads[PensionTaxRate] {
+    override def reads(json: JsValue): JsResult[PensionTaxRate] =
+      json match {
+        case JsNumber(value) => JsSuccess(PensionTaxRate(value.doubleValue()))
+        case _               => JsError("Unable to parse PensionTaxRate")
+      }
+  }
+}
